added couch_rest backend adapter
